1440: Move OB response factories to payments to v3_1_10 pkg (#344)
The existing OB response factories for payments are tied to the v3.1.10 schema, these classes can now be found in pkg: com.forgerock.sapi.gateway.ob.uk.rs.server.api.obie.payment.factories.v3_1_10

New factories will be required for v4.0.0 schemas.

https://github.com/SecureApiGateway/SecureApiGateway/issues/1440
