+++ secure-api-gateway-ob-uk-rs-server/src/test/java/com/forgerock/sapi/gateway/ob/uk/rs/server/api/obie/payment/factories/v3_1_10/OBWriteDomesticScheduledConsentResponse5FactoryTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.forgerock.sapi.gateway.ob.uk.rs.server.api.obie.payment.factories;
+package com.forgerock.sapi.gateway.ob.uk.rs.server.api.obie.payment.factories.v3_1_10;
 
 import static com.forgerock.sapi.gateway.ob.uk.rs.server.util.BeanValidationTestUtils.verifyBeanValidationIsSuccessful;
 import static org.assertj.core.api.Assertions.assertThat;
@@ -21,29 +21,29 @@ import static org.assertj.core.api.Assertions.assertThat;
 import org.joda.time.DateTime;
 import org.junit.jupiter.api.Test;
 
-import com.forgerock.sapi.gateway.ob.uk.rs.obie.api.payment.v3_1_10.domesticstandingorders.DomesticStandingOrderConsentsApi;
-import com.forgerock.sapi.gateway.ob.uk.rs.server.api.obie.payment.v3_1_10.domesticstandingorders.DomesticStandingOrderConsentsApiControllerTest;
-import com.forgerock.sapi.gateway.rcs.consent.store.datamodel.payment.domesticstandingorder.v3_1_10.DomesticStandingOrderConsent;
+import com.forgerock.sapi.gateway.ob.uk.rs.obie.api.payment.v3_1_10.domesticscheduledpayments.DomesticScheduledPaymentConsentsApi;
+import com.forgerock.sapi.gateway.ob.uk.rs.server.api.obie.payment.v3_1_10.domesticscheduledpayments.DomesticScheduledPaymentConsentsApiControllerTest;
+import com.forgerock.sapi.gateway.rcs.consent.store.datamodel.payment.domesticscheduled.v3_1_10.DomesticScheduledPaymentConsent;
 
 import uk.org.openbanking.datamodel.v3.payment.OBPaymentConsentStatus;
-import uk.org.openbanking.datamodel.v3.payment.OBWriteDomesticStandingOrderConsent5;
-import uk.org.openbanking.datamodel.v3.payment.OBWriteDomesticStandingOrderConsent5Data;
-import uk.org.openbanking.datamodel.v3.payment.OBWriteDomesticStandingOrderConsentResponse6;
-import uk.org.openbanking.datamodel.v3.payment.OBWriteDomesticStandingOrderConsentResponse6Data;
-import uk.org.openbanking.testsupport.payment.OBWriteDomesticStandingOrderConsentTestDataFactory;
+import uk.org.openbanking.datamodel.v3.payment.OBWriteDomesticScheduledConsent4;
+import uk.org.openbanking.datamodel.v3.payment.OBWriteDomesticScheduledConsent4Data;
+import uk.org.openbanking.datamodel.v3.payment.OBWriteDomesticScheduledConsentResponse5;
+import uk.org.openbanking.datamodel.v3.payment.OBWriteDomesticScheduledConsentResponse5Data;
+import uk.org.openbanking.testsupport.payment.OBWriteDomesticScheduledConsentTestDataFactory;
 
-class OBWriteDomesticStandingOrderConsentResponse6FactoryTest {
+class OBWriteDomesticScheduledConsentResponse5FactoryTest {
 
-    private final OBWriteDomesticStandingOrderConsentResponse6Factory factory = new OBWriteDomesticStandingOrderConsentResponse6Factory();
+    private final OBWriteDomesticScheduledConsentResponse5Factory factory = new OBWriteDomesticScheduledConsentResponse5Factory();
 
     @Test
     void testCreateConsentResponse() {
-        final OBWriteDomesticStandingOrderConsent5 consentRequest = OBWriteDomesticStandingOrderConsentTestDataFactory.aValidOBWriteDomesticStandingOrderConsent5();
-        final OBWriteDomesticStandingOrderConsent5Data requestData = consentRequest.getData();
-        final DomesticStandingOrderConsent consent = DomesticStandingOrderConsentsApiControllerTest.buildAwaitingAuthorisationConsent(consentRequest);
-        final OBWriteDomesticStandingOrderConsentResponse6 response = factory.buildConsentResponse(consent, DomesticStandingOrderConsentsApi.class);
+        final OBWriteDomesticScheduledConsent4 consentRequest = OBWriteDomesticScheduledConsentTestDataFactory.aValidOBWriteDomesticScheduledConsent4();
+        final OBWriteDomesticScheduledConsent4Data requestData = consentRequest.getData();
+        final DomesticScheduledPaymentConsent consent = DomesticScheduledPaymentConsentsApiControllerTest.buildAwaitingAuthorisationConsent(consentRequest);
+        final OBWriteDomesticScheduledConsentResponse5 response = factory.buildConsentResponse(consent, DomesticScheduledPaymentConsentsApi.class);
 
-        final OBWriteDomesticStandingOrderConsentResponse6Data responseData = response.getData();
+        final OBWriteDomesticScheduledConsentResponse5Data responseData = response.getData();
         assertThat(responseData.getStatus()).isEqualTo(OBPaymentConsentStatus.AWAITINGAUTHORISATION);
         assertThat(responseData.getConsentId()).isEqualTo(consent.getId());
         assertThat(responseData.getCreationDateTime()).isEqualTo(new DateTime(consent.getCreationDateTime()));
@@ -54,18 +54,19 @@ class OBWriteDomesticStandingOrderConsentResponse6FactoryTest {
         assertThat(responseData.getScASupportData()).isEqualTo(requestData.getScASupportData());
         assertThat(responseData.getReadRefundAccount()).isEqualTo(requestData.getReadRefundAccount());
         assertThat(responseData.getAuthorisation()).isEqualTo(requestData.getAuthorisation());
-        assertThat(responseData.getInitiation()).usingRecursiveComparison().isEqualTo(requestData.getInitiation());
+        assertThat(responseData.getInitiation()).isEqualTo(requestData.getInitiation());
         assertThat(responseData.getPermission()).isEqualTo(requestData.getPermission());
 
         // Not currently generating data for these optional response fields
         assertThat(responseData.getCutOffDateTime()).isNull();
+        assertThat(responseData.getExpectedExecutionDateTime()).isNull();
+        assertThat(responseData.getExpectedExecutionDateTime()).isNull();
 
         assertThat(response.getRisk()).isEqualTo(consentRequest.getRisk());
         assertThat(response.getLinks().getSelf().toString())
-                .endsWith("/open-banking/v3.1.10/pisp/domestic-standing-order-consents/" + consent.getId());
+                .endsWith("/open-banking/v3.1.10/pisp/domestic-scheduled-payment-consents/" + consent.getId());
         assertThat(response.getMeta()).isNotNull();
 
         verifyBeanValidationIsSuccessful(response);
     }
-    
 }
